+++ src/app/shared/forms/destacados-delete/destacados-delete.form.ts
@@ -1,26 +1,26 @@
-import { Component, inject, input, output } from '@angular/core';
+import { Component, inject, input, output, signal } from '@angular/core';
 import { DestacadoService } from '../../../core/services/destacado.service';
 import { ToastService } from '../../../core/services/toast.service';
 import { deleteDestacadoFolder } from '../../../core/utils/firebase-delete';
+import { ModalConfirm } from "../../ui/modal-confirm/modal-confirm";
 
 @Component({
   selector: 'form-destacados-delete',
-  imports: [],
+  imports: [ModalConfirm],
   templateUrl: './destacados-delete.form.html',
   styleUrl: './destacados-delete.form.css',
 })
 export class DestacadosDeleteForm {
   private destacadoService = inject(DestacadoService);
   private toastService = inject(ToastService);
-  object = output();
+
+  object = output<any>();
   onEdit = input<boolean>(false);
+  confirmOpen = signal(false);
+  pendingDeleteId = signal<number | null>(null);
 
   get destacados() {
     return this.destacadoService.destacadoItem();
-  }
-
-  setObject(obj: any): void {
-    return this.object.emit(obj);
   }
 
   ngOnInit() {
@@ -35,26 +35,34 @@ export class DestacadosDeleteForm {
     });
   }
 
+  setObject(obj: any): void {
+    this.object.emit(obj);
+  }
+
   getForId(id: number) {
     this.destacadoService.getDestacoForId(id).subscribe({
       next: (res) => {
-        console.log('item por id', res);
         this.setObject(res);
         this.destacadoService.scrollToEdit();
       },
       error: (err) => {
-        console.log('error al buscar por id', err);
+        console.error('error al buscar por id', err);
       },
     });
   }
 
-  async deleteDestacado(id: number) {
+  async confirmDelete() {
+    const id = this.pendingDeleteId();
+    if (id === null) return;
+
     try {
       await deleteDestacadoFolder(id);
+
       this.destacadoService.deleteDestacado(id).subscribe({
         next: () => {
           this.toastService.success('Destacado eliminado con éxito!');
           this.destacadoService.updateDestacado(id);
+          this.cancelDelete();
         },
         error: (err) => {
           this.toastService.error('ERROR al eliminar destacado');
@@ -65,5 +73,14 @@ export class DestacadosDeleteForm {
       console.error('Error eliminando imágenes del destacado', err);
       this.toastService.error('Error eliminando imágenes del destacado');
     }
+  }
+
+  requestDelete(id: number) {
+    this.pendingDeleteId.set(id);
+    this.confirmOpen.set(true);
+  }
+  cancelDelete() {
+    this.confirmOpen.set(false);
+    this.pendingDeleteId.set(null);
   }
 }
